# Create/allocate keys outside the hot path