As the title says. It should also obsolete #16.
Every string instantiation in `key(i)` caused a heap allocation, which tainted the benchmarks.

**Before**
```
$ go version
go version go1.21.4 linux/amd64
goos: linux
goarch: amd64
pkg: github.com/allegro/bigcache-bench
cpu: 12th Gen Intel(R) Core(TM) i7-12800H
BenchmarkMapSetForStruct-20                         4574           1050507 ns/op          697285 B/op      19746 allocs/op
BenchmarkSyncMapSetForStruct-20                     1512           3062503 ns/op         1828855 B/op      69662 allocs/op
BenchmarkOracamanMapSetForStruct-20                 3039           1593436 ns/op         1216733 B/op      20144 allocs/op
BenchmarkFreeCacheSetForStruct-20                   1172           4117920 ns/op         6987192 B/op      40287 allocs/op
BenchmarkBigCacheSetForStruct-20                    1362           3640569 ns/op         3771568 B/op      42459 allocs/op
BenchmarkMapSetForBytes-20                          2886           1722816 ns/op         2097688 B/op      29749 allocs/op
BenchmarkSyncMapSetForBytes-20                      1399           3438202 ns/op         3112061 B/op      79919 allocs/op
BenchmarkOracamanMapSetForBytes-20                  2138           2121852 ns/op         2943254 B/op      30175 allocs/op
BenchmarkFreeCacheSetForBytes-20                    1424           3446344 ns/op         7867129 B/op      30537 allocs/op
BenchmarkBigCacheSetForBytes-20                     1602           3034032 ns/op         4651572 B/op      32713 allocs/op
BenchmarkMapGetForStruct-20                     38352931               111.5 ns/op            23 B/op          1 allocs/op
BenchmarkSyncMapGetForStruct-20                 35730880               133.3 ns/op            23 B/op          1 allocs/op
BenchmarkOracamanMapGetForStruct-20             36561675               130.1 ns/op            23 B/op          1 allocs/op
BenchmarkFreeCacheGetForStruct-20                9007900               525.5 ns/op           271 B/op          8 allocs/op
BenchmarkBigCacheGetForStruct-20                 9274410               515.5 ns/op           287 B/op          9 allocs/op
BenchmarkMapGetForBytes-20                      42340047               111.4 ns/op            23 B/op          1 allocs/op
BenchmarkSyncMapGetForBytes-20                  35446474               135.0 ns/op            23 B/op          1 allocs/op
BenchmarkOracamanMapGetForBytes-20              34858657               135.3 ns/op            23 B/op          1 allocs/op
BenchmarkFreeCacheGetForBytes-20                22743499               207.3 ns/op           135 B/op          2 allocs/op
BenchmarkBigCacheGetForBytes-20                 25927096               186.1 ns/op           151 B/op          3 allocs/op
BenchmarkSyncMapSetParallelForStruct-20          8432505               687.5 ns/op            72 B/op          5 allocs/op
BenchmarkOracamanMapSetParallelForStruct-20     113340970               42.98 ns/op           30 B/op          2 allocs/op
BenchmarkFreeCacheSetParallelForStruct-20       84426818                54.65 ns/op           53 B/op          4 allocs/op
BenchmarkBigCacheSetParallelForStruct-20        62022183                74.63 ns/op          222 B/op          4 allocs/op
BenchmarkSyncMapSetParallelForBytes-20           7928991               647.5 ns/op           200 B/op          6 allocs/op
BenchmarkOracamanMapSetParallelForBytes-20      88186278                55.89 ns/op          143 B/op          3 allocs/op
BenchmarkFreeCacheSetParallelForBytes-20        84665589                52.74 ns/op          141 B/op          3 allocs/op
BenchmarkBigCacheSetParallelForBytes-20         57134434                90.22 ns/op          506 B/op          3 allocs/op
BenchmarkSyncMapGetParallelForStruct-20         189450904               25.36 ns/op           23 B/op          1 allocs/op
BenchmarkOracamanMapGetParallelForStruct-20     186230124               26.05 ns/op           23 B/op          1 allocs/op
BenchmarkFreeCacheGetParallelForStruct-20       45472926                99.74 ns/op          271 B/op          8 allocs/op
BenchmarkBigCacheGetParallelForStruct-20        43793809               109.5 ns/op           288 B/op          9 allocs/op
BenchmarkSyncMapGetParallelForBytes-20          206990751               23.17 ns/op           23 B/op          1 allocs/op
BenchmarkOracamanMapGetParallelForBytes-20      190093576               25.74 ns/op           23 B/op          1 allocs/op
BenchmarkFreeCacheGetParallelForBytes-20        91603290                48.53 ns/op          135 B/op          2 allocs/op
BenchmarkBigCacheGetParallelForBytes-20         90818000                47.96 ns/op          152 B/op          3 allocs/op
PASS
ok      github.com/allegro/bigcache-bench       194.972s
```

**After**
```
```
$ go version
go version go1.21.4 linux/amd64
$ go test -bench=. -benchmem -benchtime=4s ./... -timeout 30m
goos: linux
goarch: amd64
pkg: github.com/allegro/bigcache-bench
cpu: 12th Gen Intel(R) Core(TM) i7-12800H
BenchmarkMapSetForStruct-20                        10000            443928 ns/op          458776 B/op          2 allocs/op
BenchmarkSyncMapSetForStruct-20                     2472           1914924 ns/op         1589151 B/op      49915 allocs/op
BenchmarkOracamanMapSetForStruct-20                 5277            916422 ns/op          977812 B/op        399 allocs/op
BenchmarkFreeCacheSetForStruct-20                   1327           3135400 ns/op         6746175 B/op      20537 allocs/op
BenchmarkBigCacheSetForStruct-20                    2001           2537026 ns/op         3529876 B/op      22711 allocs/op
BenchmarkMapSetForBytes-20                          5002            908158 ns/op         1857320 B/op      10002 allocs/op
BenchmarkSyncMapSetForBytes-20                      1731           2525447 ns/op         2871165 B/op      60171 allocs/op
BenchmarkOracamanMapSetForBytes-20                  3711           1359587 ns/op         2702294 B/op      10427 allocs/op
BenchmarkFreeCacheSetForBytes-20                    1736           2730390 ns/op         7624457 B/op      10787 allocs/op
BenchmarkBigCacheSetForBytes-20                     2108           2174080 ns/op         4409556 B/op      12963 allocs/op
BenchmarkMapGetForStruct-20                     141019606               34.56 ns/op            0 B/op          0 allocs/op
BenchmarkSyncMapGetForStruct-20                 100000000               43.88 ns/op            0 B/op          0 allocs/op
BenchmarkOracamanMapGetForStruct-20             89178991                51.44 ns/op            0 B/op          0 allocs/op
BenchmarkFreeCacheGetForStruct-20               10634916               442.9 ns/op           264 B/op          7 allocs/op
BenchmarkBigCacheGetForStruct-20                11199600               431.9 ns/op           256 B/op          7 allocs/op
BenchmarkMapGetForBytes-20                      138045015               34.10 ns/op            0 B/op          0 allocs/op
BenchmarkSyncMapGetForBytes-20                  100000000               42.15 ns/op            0 B/op          0 allocs/op
BenchmarkOracamanMapGetForBytes-20              90671293                52.91 ns/op            0 B/op          0 allocs/op
BenchmarkFreeCacheGetForBytes-20                32486640               148.5 ns/op           136 B/op          2 allocs/op
BenchmarkBigCacheGetForBytes-20                 38845724               122.5 ns/op           128 B/op          2 allocs/op
BenchmarkSyncMapSetParallelForStruct-20          7906309               694.7 ns/op            73 B/op          5 allocs/op
BenchmarkOracamanMapSetParallelForStruct-20     116507072               42.41 ns/op           30 B/op          2 allocs/op
BenchmarkFreeCacheSetParallelForStruct-20       85833956                53.58 ns/op           52 B/op          4 allocs/op
BenchmarkBigCacheSetParallelForStruct-20        64296343                71.35 ns/op          217 B/op          4 allocs/op
BenchmarkSyncMapSetParallelForBytes-20           7700731               670.2 ns/op           200 B/op          6 allocs/op
BenchmarkOracamanMapSetParallelForBytes-20      87948224                55.05 ns/op          139 B/op          3 allocs/op
BenchmarkFreeCacheSetParallelForBytes-20        85810980                52.71 ns/op          142 B/op          3 allocs/op
BenchmarkBigCacheSetParallelForBytes-20         59678160                82.31 ns/op          491 B/op          3 allocs/op
BenchmarkSyncMapGetParallelForStruct-20         991513292                4.893 ns/op           0 B/op          0 allocs/op
BenchmarkOracamanMapGetParallelForStruct-20     501470618                9.533 ns/op           0 B/op          0 allocs/op
BenchmarkFreeCacheGetParallelForStruct-20       52761884                87.42 ns/op          264 B/op          7 allocs/op
BenchmarkBigCacheGetParallelForStruct-20        54872786                87.21 ns/op          256 B/op          7 allocs/op
BenchmarkSyncMapGetParallelForBytes-20          939485653                4.879 ns/op           0 B/op          0 allocs/op
BenchmarkOracamanMapGetParallelForBytes-20      475998799                9.818 ns/op           0 B/op          0 allocs/op
BenchmarkFreeCacheGetParallelForBytes-20        133355863               36.36 ns/op          136 B/op          2 allocs/op
BenchmarkBigCacheGetParallelForBytes-20         153783276               30.80 ns/op          128 B/op          2 allocs/op
PASS
ok      github.com/allegro/bigcache-bench       197.373s
```